requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
motor==3.3.1
pydantic==2.4.2
python-dotenv==1.0.0
aiofiles==23.2.1 